HAI-2310 Move the remaining DTO classes from application to hakemus
